Spawn vehicles after certain time but still have bugs

+++ Assets/Game/Scripts/Module/Scene/Gameplay/Road/IRoadModel.cs
@@ -28,7 +28,7 @@ namespace Croxxing.Module.Scene.Gameplay.Road
         public Vector3 DespawnerPosition { get; }
         public bool IsPlayerOnRoad { get; }
         public bool IsCurrentlyActive { get; }
-        public float SpawnRange { get; }
+        public int SpawnRange { get; }
         public float VehicleVelocity { get; }
     }
 }
